Finalizado #2

+++ src/main/java/br/com/trixmaps_v2/dao/LocationDao.java
@@ -8,7 +8,7 @@ public interface LocationDao {
 	
 	public void insert(Location location);
 	
-	public Location searchById(Integer id);
+	public Location searchById(Long id);
 	
 	public List<Location> listAll();
